smyrna: make zoom fields 'double's
This is part of a series of changes to standardize on `double` arithmetic in
Smyrna. This removes visual noise and improves accuracy. The computational cost
of `double` vs `float` arithmetic is no longer a significant factor in
contemporary environments.

+++ cmd/smyrna/smyrnadefs.h
@@ -132,7 +132,7 @@ enum {SCHEMACOUNT = 4}; // number of colors
 	double targetx;
 	double targety;
 
-	float r;
+	double r;
     } viewport_camera;
 
     typedef struct {
@@ -228,7 +228,7 @@ enum {SCHEMACOUNT = 4}; // number of colors
 
 	graph_data Graphdata;
 	float maxedgelen;
-	float fitin_zoom;
+	double fitin_zoom;
 	xdot* xDot;
 	double global_z;
 	attr_list* attributes;/*attribute list*/
@@ -249,7 +249,7 @@ enum {SCHEMACOUNT = 4}; // number of colors
 	/*view variables */
 	double panx;
 	double pany;
-	float zoom;
+	double zoom;
 
 	/*clipping coordinates, to avoid unnecesarry rendering */
 	float clipX1, clipX2, clipY1, clipY2;
